backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
django==4.2.23
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-prometheus==2.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
Faker==35.2.2
flower==2.0.1
frozenlist==1.5.0
gunicorn==23.0.0
http-ece==1.2.1
humanize==4.10.0
idna==3.10
joblib==1.4.2
kombu==5.5.4
multidict==6.1.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.2.0
psycopg==3.2.9
py-vapid==1.9.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.2
pywebpush==2.0.3
redis==6.1.1
requests==2.32.4
rest-framework-simplejwt==0.0.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tornado==6.4.2
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.15.2
